surprisingly it actually works... this time im not sure

+++ blogicum/blog/views.py
@@ -15,12 +15,16 @@ User = get_user_model()
 
 
 def category_posts(request, category_slug):
-    category = get_object_or_404(Category, slug=category_slug,
-                                 is_published=True)
+    category = get_object_or_404(
+        Category,
+        slug=category_slug,
+        is_published=True
+    )
     posts = Post.objects.filter(
         category=category,
         is_published=True,
-        pub_date__lte=timezone.now()
+        pub_date__lte=timezone.now(),
+        category__is_published=True
     ).annotate(
         comment_count=Count('comments')
     ).order_by('-pub_date')
@@ -31,7 +35,7 @@ def category_posts(request, category_slug):
     page_obj = paginator.get_page(page_number)
     return render(request, 'blog/category.html', {
         'category': category,
-        'page_obj': page_obj,
+        'page_obj': page_obj
     })
 
 
@@ -41,7 +45,7 @@ def index(request):
         pub_date__lte=timezone.now(),
         category__is_published=True
     ).annotate(
-        comment_count=Count('comments', filter=Q(comments__is_approved=True))
+        comment_count=Count('comments')
     ).order_by('-pub_date')
     paginator = Paginator(posts, 10)
     page_number = request.GET.get('page')
@@ -53,11 +57,9 @@ def index(request):
 
 def post_detail(request, post_id):
     post = get_object_or_404(Post, id=post_id)
-    if (
-        (not post.is_published or post.pub_date > timezone.now())
-        and (not request.user.is_authenticated or request.user != post.author)
-    ):
-        raise Http404("Post is not available")
+    if not post.is_published or post.pub_date > timezone.now():
+        if not request.user.is_authenticated or request.user != post.author:
+            raise Http404("Post is not available")
     comments = post.comments.all().order_by('created_at')
     form = CommentForm()
     return render(request, 'blog/detail.html', {
@@ -76,8 +78,9 @@ def register(request):
             return redirect('blog:profile', username=user.username)
     else:
         form = UserCreationForm()
-    return render(request, 'registration/registration_form.html',
-                  {'form': form})
+    return render(request, 'registration/registration_form.html', {
+        'form': form
+    })
 
 
 @login_required
@@ -86,7 +89,7 @@ def profile(request, username):
     posts = Post.objects.filter(author=profile_user).annotate(
         comment_count=Count('comments')
     ).order_by('-pub_date')
-    if not request.user.is_authenticated or request.user != profile_user:
+    if request.user != profile_user:
         posts = posts.filter(
             is_published=True,
             pub_date__lte=timezone.now(),
@@ -97,32 +100,25 @@ def profile(request, username):
     page_obj = paginator.get_page(page_number)
     return render(request, 'blog/profile.html', {
         'profile': profile_user,
-        'page_obj': page_obj,
+        'page_obj': page_obj
     })
 
 
 @login_required
 def edit_profile(request, username=None):
-    if username is None:
-        user = request.user
-    else:
-        user = get_object_or_404(User, username=username)
+    user = request.user if username is None else get_object_or_404(
+        User, username=username)
     if request.user != user:
         return redirect('blog:profile', username=user.username)
     if request.method == 'POST':
         form = ProfileForm(request.POST, instance=user)
         if form.is_valid():
-            form.save()
-            user.first_name = form.cleaned_data.get('first_name', '')
-            user.last_name = form.cleaned_data.get('last_name', '')
-            user.save()
+            user = form.save()
             return redirect('blog:profile', username=user.username)
     else:
         form = ProfileForm(instance=user)
-    posts = user.posts.filter(is_published=True).order_by('-pub_date')
     return render(request, 'blog/user.html', {
-        'user': request.user,
-        'posts': posts,
+        'user': user,
         'form': form
     })
 
@@ -150,7 +146,7 @@ def edit_post(request, post_id):
         form = PostForm(request.POST, request.FILES, instance=post,
                         user=request.user)
         if form.is_valid():
-            form.save()
+            post = form.save()
             return redirect('blog:post_detail', post_id=post.id)
     else:
         form = PostForm(instance=post, user=request.user)
@@ -163,7 +159,7 @@ def delete_post(request, post_id):
     if request.method == 'POST':
         post.delete()
         return redirect('blog:profile', username=request.user.username)
-    return redirect('blog:detail', post_id=post.id)
+    return redirect('blog:post_detail', post_id=post.id)
 
 
 @login_required
@@ -176,13 +172,17 @@ def add_comment(request, post_id):
             comment.post = post
             comment.author = request.user
             comment.save()
-            return redirect('blog:post_detail', post_id=post.id)
     return redirect('blog:post_detail', post_id=post.id)
 
 
 @login_required
 def edit_comment(request, post_id, comment_id):
-    comment = get_object_or_404(Comment, id=comment_id, author=request.user)
+    comment = get_object_or_404(
+        Comment,
+        id=comment_id,
+        author=request.user,
+        post_id=post_id
+    )
     if request.method == 'POST':
         form = CommentForm(request.POST, instance=comment)
         if form.is_valid():
@@ -198,6 +198,12 @@ def edit_comment(request, post_id, comment_id):
 
 @login_required
 def delete_comment(request, post_id, comment_id):
-    comment = get_object_or_404(Comment, id=comment_id, author=request.user)
-    comment.delete()
+    comment = get_object_or_404(
+        Comment,
+        id=comment_id,
+        author=request.user,
+        post_id=post_id
+    )
+    if request.method == 'POST':
+        comment.delete()
     return redirect('blog:post_detail', post_id=post_id)
